Example python script to update ownership of objects based on list passed
Right now tspublic/v1/update/ownership will update the ownership of all
the objects from user1 to user2 of which user1 is owner.
Introduced new optional parameter which has GUID of objects, which
are going to act as filter, it will only update the ownership of objects
present in this filter.